+++ src/main/java/br/com/rdevs/ecommerce/cadastro/model/dto/AlterarSenha.java
@@ -2,10 +2,12 @@ package br.com.rdevs.ecommerce.cadastro.model.dto;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class AlterarSenha {
 
-    private Long idCliente;
+    private BigInteger idCliente;
     private String senhaAtual;
     private String novaSenha;
     private String confirmarSenha;
