+++ src/main/java/br/com/rdevs/ecommerce/cadastro/repository/CartaoRepository.java
@@ -4,12 +4,13 @@ import br.com.rdevs.ecommerce.cadastro.model.entity.TbCartaoCredito;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface CartaoRepository extends JpaRepository<TbCartaoCredito, Long> {
+public interface CartaoRepository extends JpaRepository<TbCartaoCredito, BigInteger> {
 
-    List<TbCartaoCredito> findByClienteCartaoIdCliente(Long idCliente);
+    List<TbCartaoCredito> findByClienteCartaoIdCliente(BigInteger idCliente);
 
 
 }
