+++ src/main/java/br/com/rdevs/ecommerce/cadastro/repository/EnderecoRepository.java
@@ -4,10 +4,11 @@ import br.com.rdevs.ecommerce.cadastro.model.entity.TbEndereco;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 
 
 @Repository
-public interface EnderecoRepository extends JpaRepository<TbEndereco,Long> {
+public interface EnderecoRepository extends JpaRepository<TbEndereco, BigInteger> {
 
 
 }
