+++ src/main/java/br/com/rdevs/ecommerce/documentoFiscal/model/dto/PostDocumentoFiscalItemDTO.java
@@ -2,13 +2,15 @@ package br.com.rdevs.ecommerce.documentoFiscal.model.dto;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class PostDocumentoFiscalItemDTO {
 
 
-    private Long qtProduto;
+    private Integer qtProduto;
 
-    private Long cdProduto;
+    private BigInteger cdProduto;
