+++ src/main/java/br/com/rdevs/ecommerce/estoque/controller/FilialController.java
@@ -8,6 +8,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigInteger;
+
 @RestController
 public class FilialController {
 
@@ -15,7 +17,7 @@ public class FilialController {
         private FilialService service;
 
         @GetMapping("/filiais/{cdFilial}")
-        public ResponseEntity buscarTodos(@PathVariable("cdFilial") Long cdFilial){
+        public ResponseEntity buscarTodos(@PathVariable("cdFilial") BigInteger cdFilial){
             return ResponseEntity.ok().body(service.buscarPorFilial(cdFilial));
         }
