+++ src/main/java/br/com/rdevs/ecommerce/estoque/model/dto/EstoqueFilialDTO.java
@@ -2,8 +2,10 @@ package br.com.rdevs.ecommerce.estoque.model.dto;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class EstoqueFilialDTO {
-    private Long cdFilial;
-    private Long cdProduto;
+    private BigInteger cdFilial;
+    private BigInteger cdProduto;
 }
