+++ src/main/java/br/com/rdevs/ecommerce/estoque/model/dto/EstoqueProdutoDTO.java
@@ -2,12 +2,15 @@ package br.com.rdevs.ecommerce.estoque.model.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 @Data
 public class EstoqueProdutoDTO {
 
-    private Long cdFilial;
+    private BigInteger cdFilial;
 //    private String nmFilial;
-    private Long qtEstoque;
-    private Long qtEmpenho;
+    private Integer qtEstoque;
+    private Integer qtEmpenho;
 
 }
