+++ src/main/java/br/com/rdevs/ecommerce/estoque/repository/EstoqueRepository.java
@@ -6,14 +6,15 @@ import br.com.rdevs.ecommerce.estoque.model.entity.TbProdutoFilialEstoque;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface EstoqueRepository extends JpaRepository<TbProdutoFilialEstoque, Long> {
-  List<TbProdutoFilialEstoque> findByProdutoFilialCdProduto(Long cdProduto);
+public interface EstoqueRepository extends JpaRepository<TbProdutoFilialEstoque, BigInteger> {
+  List<TbProdutoFilialEstoque> findByProdutoFilialCdProduto(BigInteger cdProduto);
 
-  TbProdutoFilialEstoque findByCdFilial(Long cdFilial);
+  TbProdutoFilialEstoque findByCdFilial(BigInteger cdFilial);
 
-  TbProdutoFilialEstoque findByProdutoFilialCdProdutoAndCdFilial(Long cdProduto,Long cdFilial);
+  TbProdutoFilialEstoque findByProdutoFilialCdProdutoAndCdFilial(BigInteger cdProduto,BigInteger cdFilial);
 
 }
